get daily total VX

+++ src/main/java/org/vite/dex/mm/reward/RewardKeeper.java
@@ -61,7 +61,7 @@ public class RewardKeeper implements IOrderEventHandleAware {
      * @throws IOException
      */
     public Map<String, Map<Integer, BigDecimal>> calcAddressMarketReward(OrderBooks books, BlockEventStream stream,
-            double dailyReleasedVX, long startTime, long endTime) throws IOException {
+            BigDecimal dailyReleasedVX, long startTime, long endTime) throws IOException {
 
         Map<String, Map<Integer, BigDecimal>> finalRes = Maps.newHashMap(); // <Address, Map<MarketId,RewardMarket>>
         Map<Integer, RewardMarket> marketRewards = new HashMap<>(); // <MarketId, RewardMarket>
